+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 5.5.0
 * Add `scaExemption` to Transaction sale
 * Deprecate `deviceSessionId` and `fraudMerchantId` in `CreditCardGateway`, `CustomerGateway`, `PaymentMethodGateway`, and `TransactionGateway` classes
 * Add `installments` to Transaction sale